+++ packages/fxa-content-server/app/scripts/lib/fxa-client.js
@@ -214,6 +214,9 @@ define(function (require, exports, module) {
      *   @param {String} [options.sessionTokenContext] - The context for which
      *                   the session token is being created. Defaults to the
      *                   relier's context.
+     *   @param {Boolean} [options.skipCaseError] - if set to true, INCCORECT_EMAIL_CASE
+     *                   errors will be returned to be handled locally instead of automatically
+     *                   being retried in the fxa-js-client.
      *   @param {String} [options.unblockCode] - Unblock code.
      * @returns {Promise}
      */
@@ -241,6 +244,10 @@ define(function (require, exports, module) {
 
       if (options.resume) {
         signInOptions.resume = options.resume;
+      }
+
+      if (options.skipCaseError) {
+        signInOptions.skipCaseError = options.skipCaseError;
       }
 
       setMetricsContext(signInOptions, options);
@@ -609,4 +616,3 @@ define(function (require, exports, module) {
     }
   }
 });
-