+++ packages/fxa-content-server/app/tests/spec/models/account.js
@@ -199,26 +199,31 @@ define(function (require, exports, module) {
         });
       });
 
-      describe('account has a sessionToken', function () {
-        var resp;
+      describe('account has a sessionToken', () => {
+        let resp;
+        const CANONICAL_EMAIL = 'testuser@testuser.com';
 
-        beforeEach(function () {
-          account.set('sessionToken', 'session token');
-
-          sinon.stub(fxaClient, 'recoveryEmailStatus', function () {
-            return p({
-              verified: true
-            });
+        beforeEach(() => {
+          account.set({
+            email: CANONICAL_EMAIL.toUpperCase(),
+            sessionToken: 'session token'
           });
 
+          sinon.stub(fxaClient, 'recoveryEmailStatus', () => p({
+            email: CANONICAL_EMAIL,
+            verified: true
+          }));
+
           return account.sessionStatus()
-            .then(function (_resp) {
+            .then((_resp) => {
               resp = _resp;
             });
         });
 
-        it('resolves with the session information', function () {
-          assert.isTrue(resp.verified);
+        it('resolves with the session information, updates the model', () => {
+          assert.deepEqual(resp, { email: CANONICAL_EMAIL, verified: true });
+          assert.equal(account.get('email'), CANONICAL_EMAIL);
+          assert.isTrue(account.get('verified'));
         });
       });
     });
@@ -237,7 +242,6 @@ define(function (require, exports, module) {
 
         it('polls until /recovery_email/status returns `verified: true`', () => {
           assert.equal(account.sessionStatus.callCount, 3);
-          assert.isTrue(account.get('verified'));
         });
       });
 
@@ -329,6 +333,7 @@ define(function (require, exports, module) {
               },
               reason: SignInReasons.SIGN_IN,
               resume: 'resume token',
+              skipCaseError: true,
               unblockCode: 'unblock code'
             }));
           });
@@ -397,6 +402,7 @@ define(function (require, exports, module) {
               },
               reason: SignInReasons.SIGN_IN,
               resume: 'resume token',
+              skipCaseError: true,
               unblockCode: 'unblock code'
             }));
           });
@@ -404,6 +410,51 @@ define(function (require, exports, module) {
           it('updates the account with the returned data', function () {
             assert.isTrue(account.get('verified'));
             assert.equal(account.get('sessionToken'), SESSION_TOKEN);
+          });
+        });
+
+        describe('INCORRECT_EMAIL_CASE', () => {
+          let upperCaseEmail;
+
+          beforeEach(function () {
+            upperCaseEmail = EMAIL.toUpperCase();
+
+            sinon.stub(fxaClient, 'signIn', () => {
+              if (fxaClient.signIn.callCount === 1) {
+                let err = AuthErrors.toError('INCORRECT_EMAIL_CASE');
+                err.email = EMAIL;
+                return p.reject(err);
+              } else {
+                return p({});
+              }
+            });
+
+            account.set('email', upperCaseEmail);
+            return account.signIn(PASSWORD, relier, {
+              resume: 'resume token',
+              unblockCode: 'unblock code'
+            });
+          });
+
+          it('re-tries with the normalized email, updates model with normalized email', function () {
+            const expectedOptions = {
+              metricsContext: {
+                baz: 'qux',
+                foo: 'bar'
+              },
+              reason: SignInReasons.SIGN_IN,
+              resume: 'resume token',
+              skipCaseError: true,
+              unblockCode: 'unblock code'
+            };
+
+            assert.equal(fxaClient.signIn.callCount, 2);
+            assert.isTrue(
+              fxaClient.signIn.calledWith(upperCaseEmail, PASSWORD, relier, expectedOptions));
+            assert.isTrue(
+                fxaClient.signIn.calledWith(EMAIL, PASSWORD, relier, expectedOptions));
+
+            assert.equal(account.get('email'), EMAIL);
           });
         });