+++ packages/fxa-content-server/tests/functional/settings.js
@@ -20,12 +20,14 @@ define([
   var click = FunctionalHelpers.click;
   var closeCurrentWindow = FunctionalHelpers.closeCurrentWindow;
   var createUser = FunctionalHelpers.createUser;
+  var denormalizeStoredEmail = FunctionalHelpers.denormalizeStoredEmail;
   var fillOutSignIn = FunctionalHelpers.fillOutSignIn;
   var focus = FunctionalHelpers.focus;
   var getFxaClient = FunctionalHelpers.getFxaClient;
   var openPage = FunctionalHelpers.openPage;
   var openSettingsInNewTab = thenify(FunctionalHelpers.openSettingsInNewTab);
   var testElementExists = FunctionalHelpers.testElementExists;
+  var testElementTextEquals = FunctionalHelpers.testElementTextEquals;
   var testErrorTextInclude = FunctionalHelpers.testErrorTextInclude;
 
   var FIRST_PASSWORD = 'password';
@@ -94,6 +96,32 @@ define([
         // success is going to the signin page
         .findById('fxa-signin-header')
         .end();
+    },
+
+    'sign in with incorrect email case, go to settings, canonical email form is used': function () {
+      return this.remote
+        .then(openPage(SIGNIN_URL, '#fxa-signin-header'))
+        .then(fillOutSignIn(email.toUpperCase(), FIRST_PASSWORD))
+
+        .then(testElementExists('#fxa-settings-header'))
+        .then(testElementTextEquals('.card-header', email));
+    },
+
+    'sign in with incorrect email case before normalization fix, go to settings, canonical email form is used': function () {
+      return this.remote
+        .then(openPage(SIGNIN_URL, '#fxa-signin-header'))
+        .then(fillOutSignIn(email, FIRST_PASSWORD))
+
+        .then(testElementExists('#fxa-settings-header'))
+
+        // synthesize signin pre-#4470 with incorrect email case
+        .then(denormalizeStoredEmail(email))
+
+        // now, refresh to ensure the email is normalized
+        .refresh()
+
+        .then(testElementExists('#fxa-settings-header'))
+        .then(testElementTextEquals('.card-header', email));
     },
 
     'sign in, go to settings with setting param set to avatar redirects to avatar change page ': function () {